重构项目结构
