+++ MANIFEST.in
@@ -1,0 +1,2 @@
+exclude tests/*
+recursive-include src/bot/web *
