+++ run.bat
@@ -1,4 +1,4 @@
 rem @echo off
 cd  %~dp0%
-uv run -m bot.chat_app
+uv run src/bot/chat_app.py
 pause
