+++ src/bot/agent/metadata_tools.py
