+++ src/bot/web/chat_app.html
